rework rate limiting to allow bursts

+++ config.go
@@ -14,7 +14,7 @@ type Config struct {
 	PayloadUUID         string        `required:"true"`
 	PayloadOrganization string        `required:"true"`
 	ProxyHeaders        bool          `default:"false"`
-	DeliverRate         float64       `default:"2"`  // deliver requests per minute
-	FileRate            float64       `default:"10"` // file requests per minute
+	DeliverRate         int           `default:"2"`  // deliver requests per minute
+	FileRate            int           `default:"10"` // file requests per minute
 	ListenAddr          string        `default:":80"`
 }
